+++ Deep_Tryhard_DFDC/Face_detection.py
@@ -13,249 +13,12 @@ from tensorflow.keras.layers import Conv2D, Input, ZeroPadding2D, Dense, Lambda
 from tensorflow.keras.models import Model
 from tensorflow.keras.applications.mobilenet_v2 import MobileNetV2
 import tensorflow as tf
-from Deep_Tryhard_DFDC import params
+from Deep_Tryhard_DFDC.params import *
+from Deep_Tryhard_DFDC.utils import *
 tf.compat.v1.disable_eager_execution()
 
-def load_mobilenetv2_224_075_detector(path):
-    """
-    MobileNetV2 is a convolutional neural network architecture that seeks to perform well on mobile devices.
-    It is based on an inverted residual structure where the residual connections are between the bottleneck layers.
 
-    Layer description:
-    -------------------------------
-    The intermediate expansion layer uses lightweight depthwise convolutions to filter features
-    as a source of non-linearity. As a whole, the architecture of MobileNetV2 contains the initial
-    fully convolution layer with 32 filters, followed by 19 residual bottleneck layers.
-    """
-    input_tensor = Input(shape=(224, 224, 3))
-    output_tensor = MobileNetV2(weights=None, include_top=False, input_tensor=input_tensor, alpha=0.75).output
-    output_tensor = ZeroPadding2D()(output_tensor)
-    output_tensor = Conv2D(kernel_size=(3, 3), filters=5)(output_tensor)
-    model = Model(inputs=input_tensor, outputs=output_tensor)
-    model.load_weights(path)
-    return model
-
-#Pre processing
-# Converts A:B aspect rate to B:A
-def transpose_shots(shots):
-    """
-    List of aspect ratios that images could be
-    """
-    return [(shot[1], shot[0], shot[3], shot[2], shot[4]) for shot in shots]
-
-#That constant describe pieces for 16:9 images
-SHOTS = {
-    # fast less accurate
-    '2-16/9' : {
-        'aspect_ratio' : 16/9,
-        'shots' : [
-             (0, 0, 9/16, 1, 1),
-             (7/16, 0, 9/16, 1, 1)
-        ]
-    },
-    # slower more accurate
-    '10-16/9' : {
-        'aspect_ratio' : 16/9,
-        'shots' : [
-             (0, 0, 9/16, 1, 1),
-             (7/16, 0, 9/16, 1, 1),
-             (0, 0, 5/16, 5/9, 0.5),
-             (0, 4/9, 5/16, 5/9, 0.5),
-             (11/48, 0, 5/16, 5/9, 0.5),
-             (11/48, 4/9, 5/16, 5/9, 0.5),
-             (22/48, 0, 5/16, 5/9, 0.5),
-             (22/48, 4/9, 5/16, 5/9, 0.5),
-             (11/16, 0, 5/16, 5/9, 0.5),
-             (11/16, 4/9, 5/16, 5/9, 0.5),
-        ]
-    }
-}
-
-# 9:16 respectively
-SHOTS_T = {
-    '2-9/16' : {
-        'aspect_ratio' : 9/16,
-        'shots' : transpose_shots(SHOTS['2-16/9']['shots'])
-    },
-    '10-9/16' : {
-        'aspect_ratio' : 9/16,
-        'shots' : transpose_shots(SHOTS['10-16/9']['shots'])
-    }
-}
-
-def r(x):
-    return int(round(x))
-
-def sigmoid(x):
-    return 1 / (np.exp(-x) + 1)
-
-def non_max_suppression(boxes, p, iou_threshold):
-    if len(boxes) == 0:
-        return np.array([])
-
-    x1 = boxes[:, 0]
-    y1 = boxes[:, 1]
-    x2 = boxes[:, 2]
-    y2 = boxes[:, 3]
-
-    indexes = np.argsort(p)
-    true_boxes_indexes = []
-
-    while len(indexes) > 0:
-        true_boxes_indexes.append(indexes[-1])
-
-        intersection = np.maximum(np.minimum(x2[indexes[:-1]], x2[indexes[-1]]) - np.maximum(x1[indexes[:-1]], x1[indexes[-1]]), 0) * np.maximum(np.minimum(y2[indexes[:-1]], y2[indexes[-1]]) - np.maximum(y1[indexes[:-1]], y1[indexes[-1]]), 0)
-        iou = intersection / ((x2[indexes[:-1]] - x1[indexes[:-1]]) * (y2[indexes[:-1]] - y1[indexes[:-1]]) + (x2[indexes[-1]] - x1[indexes[-1]]) * (y2[indexes[-1]] - y1[indexes[-1]]) - intersection)
-
-        indexes = np.delete(indexes, -1)
-        indexes = np.delete(indexes, np.where(iou >= iou_threshold)[0])
-
-    return boxes[true_boxes_indexes]
-
-def union_suppression(boxes, threshold):
-    if len(boxes) == 0:
-        return np.array([])
-
-    x1 = boxes[:, 0]
-    y1 = boxes[:, 1]
-    x2 = boxes[:, 2]
-    y2 = boxes[:, 3]
-
-    indexes = np.argsort((x2 - x1) * (y2 - y1))
-    result_boxes = []
-
-    while len(indexes) > 0:
-        intersection = np.maximum(np.minimum(x2[indexes[:-1]], x2[indexes[-1]]) - np.maximum(x1[indexes[:-1]], x1[indexes[-1]]), 0) * np.maximum(np.minimum(y2[indexes[:-1]], y2[indexes[-1]]) - np.maximum(y1[indexes[:-1]], y1[indexes[-1]]), 0)
-        min_s = np.minimum((x2[indexes[:-1]] - x1[indexes[:-1]]) * (y2[indexes[:-1]] - y1[indexes[:-1]]), (x2[indexes[-1]] - x1[indexes[-1]]) * (y2[indexes[-1]] - y1[indexes[-1]]))
-        ioms = intersection / (min_s + 1e-9)
-        neighbours = np.where(ioms >= threshold)[0]
-        if len(neighbours) > 0:
-            result_boxes.append([min(np.min(x1[indexes[neighbours]]), x1[indexes[-1]]), min(np.min(y1[indexes[neighbours]]), y1[indexes[-1]]), max(np.max(x2[indexes[neighbours]]), x2[indexes[-1]]), max(np.max(y2[indexes[neighbours]]), y2[indexes[-1]])])
-        else:
-            result_boxes.append([x1[indexes[-1]], y1[indexes[-1]], x2[indexes[-1]], y2[indexes[-1]]])
-
-        indexes = np.delete(indexes, -1)
-        indexes = np.delete(indexes, neighbours)
-
-    return result_boxes
-
-def get_boxes_points(boxes, frame_shape):
-    result = []
-    for box in boxes:
-        lx = int(round(box[0] * frame_shape[1]))
-        ly = int(round(box[1] * frame_shape[0]))
-        rx = int(round(box[2] * frame_shape[1]))
-        ry = int(round(box[3] * frame_shape[0]))
-        result.append((lx, ly, rx, ry))
-    return result
-
-# some helper functions
-def image_resize(image, width = None, height = None, inter = cv2.INTER_AREA):
-    # initialize the dimensions of the image to be resized and
-    # grab the image size
-    dim = None
-    (h, w) = image.shape[:2]
-
-    # if both the width and height are None, then return the
-    # original image
-    if width is None and height is None:
-        return image
-
-    # check to see if the width is None
-    if width is None:
-        # calculate the ratio of the height and construct the
-        # dimensions
-        r = height / float(h)
-        dim = (int(w * r), height)
-
-    # otherwise, the height is None
-    else:
-        # calculate the ratio of the width and construct the
-        # dimensions
-        r = width / float(w)
-        dim = (width, int(h * r))
-
-    # resize the image
-    resized = cv2.resize(image, dim, interpolation = inter)
-
-    # return the resized image
-    return resized
-
-def scale_boxes(boxes, scale_w, scale_h):
-    sb = []
-    for b in boxes:
-        sb.append((b[0] * scale_w, b[1] * scale_h, b[2] * scale_w, b[3] * scale_h))
-    return sb
-
-# # draw each face separately
-# def draw_faces(filename, boxes):
-
-#     # plot each face as a subplot
-#     for i in range(len(boxes)):
-#         # get coordinates
-#         x1, y1, width, height = boxes[i]
-#         x2, y2 = x1 + width, y1 + height
-#         # define subplot
-#         plt.subplot(1, len(result_list), i+1)
-#         plt.axis('off')
-#         # plot face
-#         plt.imshow(filename[y1:y2, x1:x2])
-#     # show the plot
-#     plt.show()
-
-def from_video_to_faces(VIDEO_PATH):
-    '''
-    input: video_path - path for video
-    process:
-    1. perform a video capture from the video
-    2. read the image
-    3. detect the face
-    4. return the face cropped
-    '''
-
-    count = 0
-    list_video_processed = []
-    capture_image = cv2.VideoCapture(video_path)
-    while capture_image.isOpened() and count<271:
-        ret, frame = capture_image.read()
-#         fig = plt.figure(figsize=(6,6))
-#         ax = fig.add_subplot(111)
-        frame = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)
-        if ret:
-#             ax.imshow(frame)
-            boxes = detector.detect(frame)
-            # lets's draw boxes, just multiply each predicted [0, 1] relative coordinate to image side in pixels respectively
-            for box in boxes:
-                lx = int(round(box[0] * frame.shape[1]))
-                ly = int(round(box[1] * frame.shape[0]))
-                rx = int(round(box[2] * frame.shape[1]))
-                ry = int(round(box[3] * frame.shape[0]))
-                # x, y, w, h here
-#                 ax.imshow(frame[ly:ry, lx:rx])
-#                 imwriteframe[ly:ry, lx:rx]
-                # RESIZE
-#                 scale_percent = 120 # percent of original size
-#                 width = int(frame[ly:ry, lx:rx].shape[1] * scale_percent / 100)
-#                 height = int(frame[ly:ry, lx:rx].shape[0] * scale_percent / 100)
-#                 dim = (width, height)
-#                 resized = cv2.resize(frame[ly:ry, lx:rx], dim, interpolation = cv2.INTER_AREA)
-
-#                 ax.imshow(resized)
-#                 print(rx-lx, ry-ly)
-#                 ax.add_patch(Rectangle((lx,ly),rx - lx,ry - ly,linewidth=2,edgecolor='r',facecolor='none'))
-                list_video_processed.append(os.path.basename(video_path).replace(".mp4", ""))
-                cv2.imwrite(os.path.join(PATH , os.path.basename(video_path).replace(".mp4", "")+str(count)+'.jpg'), frame[ly:ry, lx:rx])
-            count += 30 # i.e. at 30 fps, this advances one second
-            capture_image.set(cv2.CAP_PROP_POS_FRAMES, count)
-
-        else:
-            capture_image.release()
-            break
-
-def get_meta_from_json(path):
-    df = pd.read_json(os.path.join(DATA_FOLDER, path, json_file))
-    df = df.T
-    return df
+### FACE DETECTION CLASS ###
 
 class FaceDetector():
     """
@@ -277,7 +40,7 @@ class FaceDetector():
     frame - (1920, 1080, 3) or (1080, 1920, 3) RGB Image
     returns: list of 4 element tuples (left corner x, left corner y, right corner x, right corner y) of detected boxes within [0, 1] range (see box draw code below)
     """
-    def __init__(self, model=mobilenetv2, shots=[SHOTS['10-16/9'], SHOTS_T['10-9/16']], image_size=224, grids=7, iou_threshold=0.2, union_threshold=0.1):
+    def __init__(self, model=MODEL_MOBILENETV2, shots=[SHOTS['10-16/9'], SHOTS_T['10-9/16']], image_size=224, grids=7, iou_threshold=0.2, union_threshold=0.1):
         self.model = model
         self.shots = shots
         self.image_size = image_size
@@ -377,23 +140,86 @@ class FaceDetector():
         return list(boxes)
 
 
+### PERFORM A FULL TRANSFORMATION FROM VIDEO TO CROPPED DETECTED FACES ###
+def from_video_to_faces(video_path, writing_path, framerate):
+    '''
+    __init__ parameters:
+    -------------------------------
+    video_path - path directly to the video - string
+    writing_path - path where the cropped detected face will be saved - string
+    framerate - delay you want between each frame /!\ First frame is always selected. /!\
+    For exemple --- framerate = 30 means you try to detect a face every 30 frames
+
+    __process__:
+    -------------------------------
+    1. perform a video capture from the video
+    2. read the image
+    3. apply the yolo model
+    4. saved the detected face
+
+    return : cropped pictured saved to the specified path
+    '''
+
+    count = 0
+    capture_image = cv2.VideoCapture(video_path)
+    while capture_image.isOpened() and count<271:
+        ret, frame = capture_image.read()
+        if ret:
+            frame = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)
+            boxes = detector.detect(frame)
+            # lets's draw boxes, just multiply each predicted [0, 1] relative coordinate to image side in pixels respectively
+            for box in boxes:
+                lx = int(round(box[0] * frame.shape[1]))
+                ly = int(round(box[1] * frame.shape[0]))
+                rx = int(round(box[2] * frame.shape[1]))
+                ry = int(round(box[3] * frame.shape[0]))
+                cv2.imwrite(os.path.join(writing_path , os.path.basename(video_path).replace(".mp4", "")+str(count)+'.jpg'), frame[ly:ry, lx:rx])
+            count += framerate # i.e. at 30 fps, this advances one second
+            capture_image.set(cv2.CAP_PROP_POS_FRAMES, count)
+
+        else:
+            capture_image.release()
+            break
+    capture_image.release()
+    cv2.destroyAllWindows()
+
 if __name__ == "__main__":
 
+    # Initialize list of videos
+    fake_train_sample_video = os.listdir('Data/fake')[0:1000]
+    real_train_sample_video = os.listdir('Data/real')[0:1000]
+
     # Load model
-    mobilenetv2 = load_mobilenetv2_224_075_detector("../Deepfake/facedetection-mobilenetv2-size224-alpha0.75.h5")
+    mobilenetv2 = load_mobilenetv2_224_075_detector(MODEL_MOBILENETV2)
 
     # Initialize FaceDetector
     detector = FaceDetector()
 
-    # Initialize training list and json metadata
-    train_list = list(os.listdir(os.path.join(DATA_FOLDER, TRAIN_SAMPLE_FOLDER)))
-    json_file = [file for file in train_list if  file.endswith('json')][0]
+    # Initialize logs
+    fail=0
+    success=0
+    total=0
+    log=[]
 
-    # Training Dataframe
-    meta_train_df = get_meta_from_json(TRAIN_SAMPLE_FOLDER)
-    meta_train_df.loc[meta_train_df.label=='FAKE']
-    fake_train_sample_video = list(meta_train_df.loc[meta_train_df.label=='FAKE'].sample(2).index)
-
-    # Initialize loop of face detection and cropping
+    # Loop over the selected list
     for video_file in fake_train_sample_video:
-        from_video_to_faces(os.path.join(DATA_FOLDER, TRAIN_SAMPLE_FOLDER, video_file))
+        try:
+            from_video_to_faces(os.path.join(TRAIN_FOLDER_REAL, video_file), writing_path=PATH_WRITE_REAL, framerate=30)
+            success+=1
+            total+=1
+        except:
+            log.append(video_file)
+            fail+=1
+            total+=1
+            pass
+
+    # Save to specific
+    text_file = open("log_success_1000.txt", "w")
+    n = text_file.write('\n'.join(log))
+    text_file.close()
+
+    print(f'Il y a {fail} vidéos fail sur un total de {total} vidéos soit {(fail/total)*100}% de fail')
+    print(f'Il y a {success} vidéos qui fonctionnent sur un total de {total} vidéos soit {(success/total)*100}% de réussies')
+
+    # Fast resize
+    fast_resize(path=PATH_WRITE_REAL, width=224, height=224)
