+++ Deep_Tryhard_DFDC/params.py
@@ -46,22 +46,19 @@ MODEL_VERSION = 'v2'
 
 ### FACE DETECTION PARAMS (TO UPDATE)
 
-VIDEO_PATH = 'BUCKET'
+MODEL_MOBILENETV2 = 'yolo_model/facedetection-mobilenetv2-size224-alpha0.75.h5'
 
-PATH = 'BUCKET'
+TRAIN_FOLDER_FAKE = 'Data/fake'
 
-MODEL_MOBILENETV2 = 'BUCKET'
-
-DATA_FOLDER = '/mnt/d/deepfake-detection-challenge'
-
-TRAIN_SAMPLE_FOLDER = 'train_sample_videos'
+TRAIN_FOLDER_REAL = 'Data/real'
 
 TEST_FOLDER = 'test_videos'
 
-VIDEOS_DIR = '/mnt/d/deepfake-detection-challenge/train_sample_videos/'
+SAMPLE_REAL = f'{TRAIN_FOLDER_FAKE}/abarnvbtwb.mp4'
 
-SAMPLE_REAL = f'{VIDEOS_DIR}/abarnvbtwb.mp4'
+SAMPLE_FAKE = f'{TRAIN_FOLDER_FAKE}/eepezmygaq.mp4'
 
-SAMPLE_FAKE = f'{VIDEOS_DIR}/eepezmygaq.mp4'
+# title of folder as follow : images_{number of videos}vid_{number of theorical frames}tf /fake or real
+PATH_WRITE_FAKE ='Data/images_50vid_100tf/fake'
 
-path='/mnt/d/deepfake-detection-challenge/test_videos/test/'
+PATH_WRITE_REAL = 'Data/images_50vid_100tf/real'
